+++ BUILD.bazel
@@ -45,6 +45,12 @@ java_binary(
     ],
 )
 
+# Request dumper binary, which dumps the CodeGeneratorRequest to a file on disk
+# which will be identical to the one passed to the protoc-gen-java_gapic during
+# normal execution. The dumped file then can be used to run this gapic-generator
+# directly (instead of relying on protoc to start the process), which would give
+# much greater flexibility in terms of debugging features, like attaching a
+# debugger, easier work with stdout and stderr etc.
 java_binary(
     name = "protoc-gen-code_generator_request_dumper",
     main_class = "com.google.api.generator.debug.CodeGeneratorRequestDumper",
@@ -59,9 +65,11 @@ java_binary(
     ],
 )
 
+# A binary similar to protoc-gen-java_gapic but reads the CodeGeneratorRequest
+# directly from a file instead of relying on protoc to pipe it in.
 java_binary(
     name = "protoc-gen-java_gapic_from_file",
-    main_class = "com.google.api.generator.debug.MainFromFile",
+    main_class = "com.google.api.generator.debug.MainStandalone",
     runtime_deps = [
         "//src/main/java/com/google/api/generator",
         "//src/main/java/com/google/api/generator/debug",
